move phoenix functions into ctor
